added account class

+++ othhellowordmodel4-persistence/src/main/java/de/othr/im/repository/TransferRepository.java
@@ -10,9 +10,9 @@ import de.othr.im.model.MoneyTransfer;
 
 public interface TransferRepository extends JpaRepository<MoneyTransfer, Long> {
 
-	List<MoneyTransfer> findByFrom(int from);
-	List<MoneyTransfer> findByTo(int to);
-	
+	List<MoneyTransfer> findByFrom(Long from);
+	List<MoneyTransfer> findByTo(Long to);
+
 	
 	
 }
